requirements.txt
numpy>=1.26.0
pillow>=10.0.0
opencv-python-headless>=4.10.0
scikit-learn>=1.4.0
joblib>=1.3.0
